Implemented the game logic within App component
When the user clicks a card element that hasn't already been clicked,
the App component increments & updates its state variable that
contains the current score.

If the user clicks a card element that has already been clicked, the
best score becomes the value of the current score if the current score
is greater than the current best score and current score is set to 0.

Click event listeners are added to each card element when the App
component first mounts, via the useEffect hook. Whenever a card is
clicked, the card list is shuffled randomly. If a card that has already
been clicked is clicked again, the card list is reset by setting each
card's clicked variable to false and also shuffled randomly.